fix: server action
