+++ examples/05_actions/src/actions/index.ts
@@ -1,0 +1,4 @@
+'use server';
+export const greet = (name: string) => `Hello ${name} from server!`;
+
+export * from './increment.js';
