--- examples/05_actions/src/components/funcs.ts
@@ -1,15 +1,0 @@
-'use server';
-
-import type { RenderContext } from 'waku/server';
-
-export const greet = (name: string) => `Hello ${name} from server!`;
-
-// module state on server
-let counter = 0;
-
-export const getCounter = () => counter;
-
-export function increment(this: RenderContext) {
-  counter += 1;
-  this.rerender('Waku');
-}
