+++ packages/waku/src/lib/plugins/vite-plugin-rsc-transform.ts
@@ -51,12 +51,10 @@ export function rscTransformPlugin(
         return { url };
       };
       const load = async (url: string) => {
-        let source = url === id ? code : (await this.load({ id: url })).code;
+        let source = url === id ? code : (await this.load({ id: url })).code as string
         // HACK move directives before import statements.
-        source = source!.replace(
-          /^(import {.*?} from ".*?";)\s*"use (client|server)";/,
-          '"use $2";$1',
-        );
+        source.includes("\"use client\";") && (source = source.replace("\"use client\";", "")) && (source = "\"use client\";" + source)
+        source.includes("\"use server\";") && (source = source.replace("\"use server\";", "")) && (source = "\"use server\";" + source)
         return { format: 'module', source };
       };
       RSDWNodeLoader.resolve(
